backend/requirements.txt
# Core
fastapi>=0.95.2,<0.100.0
uvicorn[standard]>=0.22.0,<0.25.0
sqlalchemy>=2.0.15,<2.1.0
alembic>=1.11.2,<1.13.0
psycopg2-binary>=2.9.6,<2.10.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
email-validator>=1.3.1,<2.0.0
python-dotenv>=1.0.0,<2.0.0
redis>=4.5.5,<5.0.0
pydantic[email]>=1.10.7,<2.0.0
python-dateutil>=2.8.2,<3.0.0

# LangChain and AI/ML
langchain>=0.0.335,<0.1.0
openai>=0.28.0,<0.29.0
tiktoken>=0.5.2,<0.6.0

# Async support
anyio>=3.7.1,<4.0.0
httpx>=0.24.1,<0.26.0

# Development
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.1,<0.23.0
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.1,<2.0.0
